doc: refine long term support versioning strategy
Hoping to find the right balance here. The goal is to find a RC
and LTS maintenance versioning strategy that does not interfere
with the master versioning, allows LTS releases to be cut
organically at any time without impacting the existing versioning
scheme, and allows easier tracking of post-release changes.

I know patch metadata isn't everyone's favorite thing in the world
but it would appear to be the best overall strategy for post release
changes.